+++ src/chart/rendering/markers/connectingmarker.cpp
@@ -158,6 +158,8 @@ ConnectingMarker::ConnectingMarker(const DrawingContext &ctx,
 
 	dataRect.pos = isLine ? points[2] : points[1];
 	dataRect.size = Geom::Size{points[2] - dataRect.pos};
+
+	setDataPosition(ctx.coordSys);
 }
 
 const Gen::Marker *ConnectingMarker::getPrev(