+++ src/chart/rendering/markers/rectanglemarker.h
@@ -10,6 +10,7 @@ class RectangleMarker : public SingleDrawMarker
 {
 public:
 	RectangleMarker(const Gen::Marker &marker,
+	    const CoordinateSystem &coordSys,
 	    const Gen::Options &options,
 	    const Styles::Chart &style);
 };