+++ CHANGELOG.md
@@ -4,16 +4,25 @@
 
 ### Fixed
 
+- Do not interpolate hiding/showing legend
+
+## [0.15.0] - 2024-10-28
+
+### Fixed
+
 - Markers are the same even if new record added.
+- Flying out marker label fixed.
 - Axis line hide/show at same time with axis labels/ticks/title.
 - Do not draw invisible axis line.
-- Do not interpolate hiding/showing legend
 
 ### Changed
 
 - Removed 'min' align property from the API which equivalent with the 'none'.
 - Changed MarkerId to be a string instead of a number.
 
+### Added
+
+- Add marker top and center position to draw event.
 
 ## [0.14.0] - 2024-10-03